chore: move schemas into separate file
Easier to find, remember and manage them when there's a standard form for the schema location.
